guard/requirements.txt
Pillow>=10.0.0
httpx>=0.27.0
plyer>=2.1.0
psutil>=5.9.0
fastapi>=0.115.0
uvicorn>=0.30.0
